+++ plugin/android-junit5/src/test/groovy/de/mannodermaus/gradle/plugins/junit5/DslGroovyTests.groovy
@@ -34,7 +34,7 @@ class DslGroovyTests {
         .asAndroidApplication()
         .build()
 
-    project.android.testOptions.junitPlatform {
+    project.junitPlatform {
       filters {
         includeTags "some-tag"
         excludeTags "other-tag"